--- lib/core/ResponseNotFound.re
@@ -1,3 +1,0 @@
-let run = (_, response) => {
-    Rest.bodyResponse(response, 404, "File not Found")
-}
