--- lib/core/Rest.bs.js
@@ -1,24 +1,0 @@
-// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
-'use strict';
-
-var Http = require("bs-node-http/src/Http.js");
-
-function bodyResponse(response, statusCode, message) {
-  var arg = function (param, param$1) {
-    return Http.ServerResponse.write(message, param, param$1);
-  };
-  arg(undefined, Http.ServerResponse.setStatusCode(statusCode, response)).end();
-  
-}
-
-function bodyJsonResponse(response, statusCode, message) {
-  var arg = function (param, param$1) {
-    return Http.ServerResponse.write(message, param, param$1);
-  };
-  arg(undefined, Http.ServerResponse.setStatusCode(statusCode, Http.ServerResponse.setHeader("content-type", "application/json", response))).end();
-  
-}
-
-exports.bodyResponse = bodyResponse;
-exports.bodyJsonResponse = bodyJsonResponse;
-/* No side effect */
