--- lib/core/Routers.bs.js
@@ -1,23 +1,0 @@
-// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
-'use strict';
-
-var Http = require("bs-node-http/src/Http.js");
-var List = require("bs-platform/lib/js/list.js");
-var Curry = require("bs-platform/lib/js/curry.js");
-var Method = require("./Method.bs.js");
-
-function start(request, response, routers) {
-  var methodHttp = Method.getMethod(Http.ClientRequest.getMethod(request));
-  var uri = Http.ClientRequest.getUrl(request);
-  return List.map((function (router) {
-                var metheod = router.method;
-                var uriString = router.uri;
-                if (metheod === methodHttp && uri === uriString) {
-                  return Curry._2(router.action, request, response);
-                }
-                
-              }), routers);
-}
-
-exports.start = start;
-/* No side effect */
