+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "reason-http-node",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "private": false,
   "scripts": {
     "dev": "nodemon lib/sample/App.bs.js",
